restructured the folders and files
