--- app/dashboard/diary/layout.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-export default function DiaryLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="max-w-7xl mx-auto px-4 sm:px-6 md:px-8">
-      {children}
-    </div>
-  );
-}
