+++ app/diary/page.tsx
@@ -1,0 +1,5 @@
+import DiaryPage from '@/components/pages/DiaryPage';
+
+export default function Page() {
+  return <DiaryPage />;
+}
