+++ app/measurements/page.tsx
@@ -1,0 +1,5 @@
+import MeasurementsPage from '@/components/pages/MeasurementsPage';
+
+export default function Page() {
+  return <MeasurementsPage />;
+}
