+++ app/profile/page.tsx
@@ -1,0 +1,5 @@
+import ProfilePage from '@/components/pages/ProfilePage';
+
+export default function Page() {
+  return <ProfilePage />;
+}
