+++ app/workouts/page.tsx
@@ -1,0 +1,5 @@
+import WorkoutsPage from '@/components/pages/WorkoutsPage';
+
+export default function Page() {
+  return <WorkoutsPage />;
+}
