+++ components/pages/DiaryPage.tsx
@@ -189,7 +189,7 @@ export default function DiaryPage() {
   };
 
   return (
-    <div className="py-6">
+    <div className="max-w-7xl mx-auto px-4 sm:px-6 md:px-8 py-6">
       <h1 className="text-3xl font-bold text-center mb-8">Fitness Diary</h1>
       <div className="bg-white shadow rounded-lg p-6">
         <div className="text-center mb-4">
