+++ components/pages/HomePage.tsx
@@ -1,0 +1,160 @@
+'use client';
+
+import Link from 'next/link';
+
+// DEVELOPMENT MODE: Auth checks bypassed
+// TODO: Restore auth checks before production
+
+export default function HomePage() {
+  // Development mode - no auth checks
+
+  // Landing page with development mode banner
+  return (
+    <div className="bg-white">
+      {/* Development mode banner */}
+      {/* <div className="bg-yellow-100 border-b border-yellow-300 p-4">
+        <div className="max-w-7xl mx-auto flex items-center justify-between">
+          <p className="text-yellow-800">
+            <span className="font-bold">DEVELOPMENT MODE:</span> Auth checks bypassed for easier navigation.
+          </p>
+          <Link 
+            href="/dashboard" 
+            className="bg-yellow-500 hover:bg-yellow-600 text-white py-2 px-4 rounded-md font-medium shadow-sm"
+          >
+            Go to Dashboard →
+          </Link>
+        </div>
+      </div> */}
+      
+      <div className="relative overflow-hidden">
+        <div className="relative pt-6 pb-16 sm:pb-24">
+          {/* Header */}
+          <nav className="relative mx-auto flex max-w-7xl items-center justify-between px-4 sm:px-6">
+            <div className="flex flex-1 items-center justify-between">
+              <div className="flex flex-shrink-0 items-center">
+                <span className="text-2xl font-bold text-indigo-600">FitSage</span>
+              </div>
+              <div className="flex items-center space-x-4">
+                <Link
+                  href="/auth/login"
+                  className="inline-block rounded-md border border-transparent bg-indigo-600 px-4 py-2 text-base font-medium text-white hover:bg-indigo-700"
+                >
+                  Sign in
+                </Link>
+                <Link
+                  href="/auth/signup"
+                  className="inline-block rounded-md border border-transparent bg-white px-4 py-2 text-base font-medium text-indigo-600 hover:bg-gray-50"
+                >
+                  Sign up
+                </Link>
+              </div>
+            </div>
+          </nav>
+
+          {/* Hero section */}
+          <main className="mx-auto mt-16 max-w-7xl px-4 sm:mt-24 sm:px-6">
+            <div className="text-center">
+              <h1 className="text-4xl font-extrabold tracking-tight text-gray-900 sm:text-5xl md:text-6xl">
+                <span className="block">Track your fitness journey</span>
+                <span className="block text-indigo-600">with FitSage</span>
+              </h1>
+              <p className="mx-auto mt-3 max-w-md text-base text-gray-500 sm:text-lg md:mt-5 md:max-w-3xl md:text-xl">
+                Log every workout. Unlock AI-powered insights. Crush your fitness goals with your ultimate training partner.
+              </p>
+              <div className="mx-auto mt-5 max-w-md sm:flex sm:justify-center md:mt-8">
+                <div className="rounded-md shadow">
+                  <Link
+                    href="/auth/signup"
+                    className="flex w-full items-center justify-center rounded-md border border-transparent bg-indigo-600 px-8 py-3 text-base font-medium text-white hover:bg-indigo-700 md:py-4 md:px-10 md:text-lg"
+                  >
+                    Get started
+                  </Link>
+                </div>
+                <div className="mt-3 rounded-md shadow sm:mt-0 sm:ml-3">
+                  <Link
+                    href="/auth/login"
+                    className="flex w-full items-center justify-center rounded-md border border-transparent bg-white px-8 py-3 text-base font-medium text-indigo-600 hover:bg-gray-50 md:py-4 md:px-10 md:text-lg"
+                  >
+                    Sign in
+                  </Link>
+                </div>
+              </div>
+            </div>
+          </main>
+        </div>
+      </div>
+
+      {/* Features section */}
+      <div className="py-12 bg-gray-50">
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+          <div className="lg:text-center">
+            <h2 className="text-base text-indigo-600 font-semibold tracking-wide uppercase">Features</h2>
+            <p className="mt-2 text-3xl leading-8 font-extrabold tracking-tight text-gray-900 sm:text-4xl">
+              Everything you need to track your fitness
+            </p>
+          </div>
+
+          <div className="mt-10">
+            <dl className="space-y-10 md:space-y-0 md:grid md:grid-cols-2 md:gap-x-8 md:gap-y-10">
+              <div className="relative">
+                <dt>
+                  <div className="absolute flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                    <svg className="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
+                    </svg>
+                  </div>
+                  <p className="ml-16 text-lg leading-6 font-medium text-gray-900">AI-powered genius</p>
+                </dt>
+                <dd className="mt-2 ml-16 text-base text-gray-500">
+                  Instant, laser-sharp summaries and killer suggestions tailored just for you.
+                </dd>
+              </div>
+
+              <div className="relative">
+                <dt>
+                  <div className="absolute flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                    <svg className="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 11H5m14 0a2 2 0 012 2v6a2 2 0 01-2 2H5a2 2 0 01-2-2v-6a2 2 0 012-2m14 0V9a2 2 0 00-2-2M5 11V9a2 2 0 012-2m0 0V5a2 2 0 012-2h6a2 2 0 012 2v2M7 7h10" />
+                    </svg>
+                  </div>
+                  <p className="ml-16 text-lg leading-6 font-medium text-gray-900">Log workouts</p>
+                </dt>
+                <dd className="mt-2 ml-16 text-base text-gray-500">
+                  Record your exercises, sets, reps, and weights with ease.
+                </dd>
+              </div>
+
+              <div className="relative">
+                <dt>
+                  <div className="absolute flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                    <svg className="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5H7a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2" />
+                    </svg>
+                  </div>
+                  <p className="ml-16 text-lg leading-6 font-medium text-gray-900">Smart workout memory</p>
+                </dt>
+                <dd className="mt-2 ml-16 text-base text-gray-500">
+                  Never lose momentum. Your workout diary builds habits that stick and progress that lasts.
+                </dd>
+              </div>
+
+              <div className="relative">
+                <dt>
+                  <div className="absolute flex items-center justify-center h-12 w-12 rounded-md bg-indigo-500 text-white">
+                    <svg className="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z" />
+                    </svg>
+                  </div>
+                  <p className="ml-16 text-lg leading-6 font-medium text-gray-900">Progress visualization</p>
+                </dt>
+                <dd className="mt-2 ml-16 text-base text-gray-500">
+                  See your progress over time with intuitive charts and metrics.
+                </dd>
+              </div>
+            </dl>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
